+++ modules/document_repository/jsx/categoryForm.js
@@ -10,10 +10,10 @@ import {
 } from 'jsx/Form';
 
 /**
- * Document Upload Form
+ * Category Creation Form
  *
  * Fetches data from Loris backend and displays a form allowing
- * to upload a doc file
+ * to create a category
  *
  * @author Shen Wang
  * @version 1.0.0
@@ -77,6 +77,13 @@ class DocCategoryForm extends React.Component {
       return (<Loader/>);
     }
 
+    let disabled = true;
+    let addButton = null;
+    if (loris.userHasPermission('document_repository_categories')) {
+        disabled = false;
+        addButton = <ButtonElement label="Add Category"/>;
+    }
+
     return (
       <div className="row">
         <div className="col-md-8 col-lg-7">
@@ -91,6 +98,7 @@ class DocCategoryForm extends React.Component {
               label="Category Name"
               onUserInput={this.setFormData}
               required={true}
+              disabled={disabled}
               value={this.state.formData.categoryName}
             />
             <SelectElement
@@ -98,6 +106,7 @@ class DocCategoryForm extends React.Component {
               label="Parent"
               options={this.state.data.fieldOptions.fileCategories}
               onUserInput={this.setFormData}
+              disabled={disabled}
               hasError={false}
               value={this.state.formData.parentId}
             />
@@ -105,9 +114,10 @@ class DocCategoryForm extends React.Component {
               name="comments"
               label="Comments"
               onUserInput={this.setFormData}
+              disabled={disabled}
               value={this.state.formData.comments}
             />
-            <ButtonElement label="Add Category"/>
+            {addButton}
           </FormElement>
         </div>
       </div>